refactor(browser-extension): extract config + tab-group helpers from background.js
background.js sat at the oxlint max-lines budget, so any behavior change to it
tripped the gate. Split out two cohesive, side-effect-light units with no behavior
change: chrome.storage config reads (getConfig/getCopilotConfig) into
modules/config.js, and the OpenClaw tab-group consent-boundary helpers
(list/add/remove/isShared/isOpenClawGroupId/…) into modules/tab-groups.js.

Definitions move verbatim and are imported back, so call sites and the copilot
controller injection are unchanged. background.js drops from ~695 to ~631 code
lines, leaving room for follow-up fixes. Rig-verified behaviour-neutral (pair +
connect unchanged).
